+++ RaidHours/ConsoleCmdRaidHours.cs
@@ -1,0 +1,7 @@
+﻿namespace RaidHours
+{
+    internal class ConsoleCmdRaidHours
+    {
+        // TODO: add admin console commands for RaidHours mod
+    }
+}