+++ RaidHours/Settings.cs
@@ -1,0 +1,7 @@
+﻿namespace RaidHours
+{
+    internal class Settings
+    {
+        // TODO: add settings for admin configuration
+    }
+}